회원가입

+++ src/common/components/Header/Header.impl.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from "next/router";
 import { Header } from "./Header.interface";
 import VHeader from "./Header.view";
-import authService from "@domains/User/services/auth.service";
+import { parseJwt } from "@/domains/User/hooks";
 
 const Header: React.FC<Header.IProps> = () => {
     const router = useRouter();
@@ -11,6 +11,12 @@ const Header: React.FC<Header.IProps> = () => {
         !localStorage.getItem("userTokens") &&
         localStorage.getItem("userToken") !== undefined;
     // console.log("haveNoToken : " + haveNoToken);
+
+    let userInfo;
+    if (typeof window !== "undefined") {
+        userInfo = parseJwt(localStorage.getItem("userTokens"));
+        console.log(userInfo);
+    }
     // typeof window !== "undefined" &&
     //     console.log("nowToken : " + localStorage.getItem("userTokens"));
 
@@ -19,11 +25,13 @@ const Header: React.FC<Header.IProps> = () => {
     };
 
     const userLogout = () => {
-        alert("로그아웃 하시겠습니까?");
-        window.localStorage.clear();
-        // router.push("/");
-        // location.reload();
-        router.reload();
+        const answer = confirm("로그아웃 하시겠습니까?");
+        if (answer) {
+            window.localStorage.clear();
+            // router.push("/");
+            // location.reload();
+            router.reload();
+        }
     };
 
     return (
@@ -31,6 +39,8 @@ const Header: React.FC<Header.IProps> = () => {
             goToLogin={goToLogin}
             userLogout={userLogout}
             haveNoToken={haveNoToken}
+            userNickname={userInfo ? userInfo.nickname : ""}
+            // sessionTime={sessionTime}
         />
     );
 };
